requirements.txt
pygame==2.5.2
Flask==2.2.2
Flask-Cors==3.0.10
pymongo==4.6.1
Werkzeug==2.2.2
pillow==10.2.0